2025/02/21 完成数据库设计，实体类等创建

+++ src/main/java/com/yun/springbootinit/mapper/UserMapper.java
@@ -5,12 +5,9 @@ import com.yun.springbootinit.model.entity.User;
 
 /**
  * @author cheny
- * @description 针对表【user(用户)】的数据库操作Mapper
- * @createDate 2023-07-18 22:39:07
- * @Entity com.yupi.springbootinit.model.entity.User
  */
 public interface UserMapper extends BaseMapper<User> {
-    User queryUserById(Integer id);
+    User queryUserById(Long id);
 }
